LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/pseudo-get-computed-style-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: container(target) expected "multiply" but got "normal"

Harness Error (FAIL), message = Unhandled rejection: assert_equals: container(target) expected "multiply" but got "normal"

TIMEOUT position property of pseudo elements Test timed out
NOTRUN properties of pseudo elements outside of transition
